+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/IndentationRule.kt
@@ -47,6 +47,7 @@ import com.pinterest.ktlint.core.ast.ElementType.RBRACKET
 import com.pinterest.ktlint.core.ast.ElementType.REGULAR_STRING_PART
 import com.pinterest.ktlint.core.ast.ElementType.RPAR
 import com.pinterest.ktlint.core.ast.ElementType.SAFE_ACCESS_EXPRESSION
+import com.pinterest.ktlint.core.ast.ElementType.SECONDARY_CONSTRUCTOR
 import com.pinterest.ktlint.core.ast.ElementType.SHORT_STRING_TEMPLATE_ENTRY
 import com.pinterest.ktlint.core.ast.ElementType.STRING_TEMPLATE
 import com.pinterest.ktlint.core.ast.ElementType.SUPER_TYPE_CALL_ENTRY
@@ -75,6 +76,7 @@ import com.pinterest.ktlint.core.ast.nextLeaf
 import com.pinterest.ktlint.core.ast.nextSibling
 import com.pinterest.ktlint.core.ast.parent
 import com.pinterest.ktlint.core.ast.prevCodeLeaf
+import com.pinterest.ktlint.core.ast.prevCodeSibling
 import com.pinterest.ktlint.core.ast.prevLeaf
 import com.pinterest.ktlint.core.ast.prevSibling
 import com.pinterest.ktlint.core.ast.upsertWhitespaceAfterMe
@@ -568,6 +570,8 @@ class IndentationRule : Rule("indent"), Rule.Modifier.RestrictToRootLast {
                             ctx.ignored.add(n)
                         }
                     }
+                    FUNCTION_LITERAL ->
+                        adjustExpectedIndentInFunctionLiteral(n, ctx)
                     WHITE_SPACE ->
                         if (n.textContains('\n')) {
                             if (
@@ -765,13 +769,24 @@ class IndentationRule : Rule("indent"), Rule.Modifier.RestrictToRootLast {
     }
 
     private fun adjustExpectedIndentAfterColon(n: ASTNode, ctx: IndentContext) {
-        expectedIndent++
-        debug { "++after(COLON) -> $expectedIndent" }
-        if (n.isPartOf(FUN)) {
-            val returnType = n.nextCodeSibling()
-            ctx.exitAdjBy(returnType!!, -1)
-        } else {
-            ctx.exitAdjBy(n.treeParent, -1)
+        when {
+            n.isPartOf(FUN) -> {
+                expectedIndent++
+                debug { "++after(COLON IN FUN) -> $expectedIndent" }
+                val returnType = n.nextCodeSibling()
+                ctx.exitAdjBy(returnType!!, -1)
+            }
+            n.treeParent.isPartOf(SECONDARY_CONSTRUCTOR) -> {
+                expectedIndent++
+                debug { "++after(COLON IN CONSTRUCTOR) -> $expectedIndent" }
+                val nextCodeSibling = n.nextCodeSibling()
+                ctx.exitAdjBy(nextCodeSibling!!, -1)
+            }
+            else -> {
+                expectedIndent++
+                debug { "++after(COLON) -> $expectedIndent" }
+                ctx.exitAdjBy(n.treeParent, -1)
+            }
         }
     }
 
@@ -779,6 +794,33 @@ class IndentationRule : Rule("indent"), Rule.Modifier.RestrictToRootLast {
         expectedIndent++
         debug { "++inside(CONDITION) -> $expectedIndent" }
         ctx.exitAdjBy(n.treeParent, -1)
+    }
+
+    private fun adjustExpectedIndentInFunctionLiteral(n: ASTNode, ctx: IndentContext) {
+        require(n.elementType == FUNCTION_LITERAL)
+
+        var countNonWhiteSpaceElementsBeforeArrow = 0
+        var arrowNode: ASTNode? = null
+        var hasWhiteSpaceWithNewLine = false
+        val iterator = n.children().iterator()
+        while (iterator.hasNext()) {
+            val child = iterator.next()
+            if (child.elementType == ARROW) {
+                arrowNode = child
+                break
+            }
+            if (child.elementType == WHITE_SPACE) {
+                hasWhiteSpaceWithNewLine = hasWhiteSpaceWithNewLine || child.text.contains("\n")
+            } else {
+                countNonWhiteSpaceElementsBeforeArrow++
+            }
+        }
+
+        if (arrowNode != null && hasWhiteSpaceWithNewLine) {
+            expectedIndent++
+            debug { "++after(FUNCTION_LITERAL) -> $expectedIndent" }
+            ctx.exitAdjBy(arrowNode.prevCodeSibling()!!, -1)
+        }
     }
 
     private fun indentStringTemplate(