+++ CHANGELOG.md
@@ -6,16 +6,24 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Added
 - Use Gradle JVM toolchain with language version 8 to compile the project
+- Basic tests for CLI ([#540](https://github.com/pinterest/ktlint/issues/540))
 
 ### Fixed
+- Fix indentation of function literal ([#1247](https://github.com/pinterest/ktlint/issues/1247))
 - Fix false positive in rule spacing-between-declarations-with-annotations ([#1281](https://github.com/pinterest/ktlint/issues/1281))
 - Fix NoSuchElementException for property accessor (`trailing-comma`) ([#1280](https://github.com/pinterest/ktlint/issues/1280))
 - Fix output of reporters. Only print rules violations to stdout, everything else is printed to stderr ([#1137](https://github.com/pinterest/ktlint/issues/1137))
+- Fix ClassCastException using ktlintFormat on class with KDoc (`no-trailing-spaces`) ([#1270](https://github.com/pinterest/ktlint/issues/1270))
+- Do not remove trailing comma in annotation ([#1297](https://github.com/pinterest/ktlint/issues/1297))
+- Do not remove import which is used as markdown link in KDoc only (`no-unused-imports`) ([#1282](https://github.com/pinterest/ktlint/issues/1282))
+- Fix indentation of secondary constructor (`indent`) ([#1222](https://github.com/pinterest/ktlint/issues/1222))
+- Fix alignment of arrow when trailing comma is missing in when entry (`trailing-comma`) ([#1312](https://github.com/pinterest/ktlint/issues/1312))
 
 ### Changed
 - Update Kotlin version to `1.6.0` release
 - Add separate tasks to run tests on JDK 11 - "testOnJdk11"
 - Update Dokka to `1.6.0` release
+- Update shadow plugin to `7.1.1` release
 
 ### Removed